Create README.md
- Recommendation system (RS) plays a vital role for the online data users.
But most of the approaches based on the RS in online public networks are not reliable with
the interest of data users. In order to overcome the untrustworthiness and uncertainty of user
based RS in public network. We have proposed a Novel Secure Key Distribution based
Recommendation System (NSKD-RS) for preserving the privacy of the data users using
neural networks. Data User (DU) send request to other DU through the Trusted Third Party
(TTP). TTP will search the similar interest DUs based on the received search request. The
DUs interest is placed in terms of tag. If the DU interest matches with other DU interest then
two DU’s will become friends by using neural networks. The cloud has the collection of the
data user’s files, here the neural networks are used for the data classification based on the
data user’s interest in addition to neural networks provide the recommended users based on
the similar interest. We have proposed the secure recommendation system based key
distribution mechanism for the tag decryption as well as to have secure communication
between the DUs. The secure tag matching process is developed in order to reduce the
computation complexity as well as to have the secure data sharing between the DUs. By
using our proposed NSKD-RS the commutation as well as the communication complexity is
reduced.